refactor: sort out assets folder
