+++ packages/frontend/src/assets/img/avatar.png
