+++ packages/frontend/src/assets/img/downvote.png
