+++ packages/frontend/src/assets/img/logo-white.png
