+++ packages/frontend/src/assets/img/logo.png
