+++ packages/frontend/src/assets/img/upvote.png
