+++ packages/frontend/src/components/layout/MobileBottomNav.tsx
@@ -1,10 +1,10 @@
 import { motion } from 'framer-motion'
 import { NavLink } from 'react-router-dom'
-import { ReactComponent as AddIcon } from '@/assets/add.svg'
-import { ReactComponent as BellIcon } from '@/assets/bell.svg'
-import { ReactComponent as HomeIcon } from '@/assets/home.svg'
-import { ReactComponent as PersonCircleIcon } from '@/assets/person-circle.svg'
-import { ReactComponent as StarIcon } from '@/assets/star.svg'
+import { ReactComponent as AddIcon } from '@/assets/svg/add.svg'
+import { ReactComponent as BellIcon } from '@/assets/svg/bell.svg'
+import { ReactComponent as HomeIcon } from '@/assets/svg/home.svg'
+import { ReactComponent as PersonCircleIcon } from '@/assets/svg/person-circle.svg'
+import { ReactComponent as StarIcon } from '@/assets/svg/star.svg'
 import SignupPendingTransition from '@/components/login/SignupPendingTransition'
 import { useIsMutating } from '@tanstack/react-query'
 import { MutationKeys } from '@/constants/queryKeys'
