+++ packages/frontend/src/components/post/Post.tsx
@@ -4,13 +4,13 @@ import React, { useEffect, useMemo, useState } from 'react'
 import LinesEllipsis from 'react-lines-ellipsis'
 import { Link } from 'react-router-dom'
 import Avatar from '@/components/common/Avatar'
-import Comment from '../../assets/comment.png'
-import Downvote from '../../assets/downvote.png'
-import Upvote from '../../assets/upvote.png'
-import { useVotes } from '../../hooks/useVotes/useVotes'
+import Comment from '@/assets/img/comment.png'
+import Downvote from '@/assets/img/downvote.png'
+import Upvote from '@/assets/img/upvote.png'
+import { useVotes } from '@/hooks/useVotes/useVotes'
 import LikeAnimation from '../ui/animations/LikeAnimation'
-import useStore from '../../store/usePostStore'
-import useVoteStore from '../../store/useVoteStore'
+import useStore from '@/store/usePostStore'
+import useVoteStore from '@/store/useVoteStore'
 import VoteFailureDialog from '@/components/post/VoteFailureDialog'
 import { PostStatus } from '@/types/Post'
 import { VoteAction } from '@/types/Vote'
