+++ packages/frontend/src/components/ui/animations/LikeAnimation.tsx
@@ -1,6 +1,6 @@
 import { motion, AnimatePresence } from 'framer-motion'
-import upvoteImage from '../../../assets/upvote.png'
-import downvoteImage from '../../../assets/downvote.png'
+import upvoteImage from '@/assets/img/upvote.png'
+import downvoteImage from '@/assets/img/downvote.png'
 import { VoteAction } from '@/types/Vote'
 
 interface LikeAnimationProps {
