+++ packages/frontend/src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import './styles/main.css'
+import '@/assets/css/main.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
