+++ packages/frontend/src/layouts/AppLayout.tsx
@@ -7,13 +7,13 @@ import {
     useMatch,
     useNavigate,
 } from 'react-router-dom'
-import { ReactComponent as ArrowLeftIcon } from '@/assets/arrow-left.svg'
-import { ReactComponent as BellIcon } from '@/assets/bell.svg'
-import { ReactComponent as HomeIcon } from '@/assets/home.svg'
-import Logo from '@/assets/logo.png'
-import { ReactComponent as PersonCircleIcon } from '@/assets/person-circle.svg'
-import { ReactComponent as SearchIcon } from '@/assets/search.svg'
-import { ReactComponent as StarIcon } from '@/assets/star.svg'
+import { ReactComponent as ArrowLeftIcon } from '@/assets/svg/arrow-left.svg'
+import { ReactComponent as BellIcon } from '@/assets/svg/bell.svg'
+import { ReactComponent as HomeIcon } from '@/assets/svg/home.svg'
+import Logo from '@/assets/img/logo.png'
+import { ReactComponent as PersonCircleIcon } from '@/assets/svg/person-circle.svg'
+import { ReactComponent as SearchIcon } from '@/assets/svg/search.svg'
+import { ReactComponent as StarIcon } from '@/assets/svg/star.svg'
 import ActionNotification from '@/components/layout/ActionNotification'
 import EpochInfo from '@/components/layout/EpochInfo'
 import MobileBottomNav from '@/components/layout/MobileBottomNav'
